Move use-fee to its own file
